requirements.txt
python-telegram-bot
numpy
python-binance
